backend/requirements.txt
setuptools>=68.0.0
fastapi==0.109.0
uvicorn[standard]==0.27.0
pymilvus==2.6.4
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.26.0
aiofiles>=23.2.1
openai>=1.0.0
tqdm>=4.65.0

sentence-transformers>=2.2.0

pyyaml>=6.0
watchfiles>=0.20.0

pytest==8.0.0
pytest-asyncio==0.23.6
